+++ airbyte-config/models/src/main/java/io/airbyte/config/Configs.java
@@ -78,9 +78,13 @@ public interface Configs {
 
   WorkerEnvironment getWorkerEnvironment();
 
+  String getSpecCacheBucket();
+
   WorkspaceRetentionConfig getWorkspaceRetentionConfig();
 
   List<WorkerPodToleration> getWorkerPodTolerations();
+
+  MaxWorkersConfig getMaxWorkers();
 
   String getTemporalHost();