+++ airbyte-integrations/bases/base-normalization/integration_tests/normalization_test_output/redshift/test_primary_key_streams/final/airbyte_tables/test_normalization/nested_stream_with_complex_columns_resulting_into_long_names.sql
@@ -15,5 +15,5 @@ select
     _airbyte_nested_stream_with_complex_columns_resulting_into_long_names_hashid
 from "integrationtests".test_normalization."nested_stream_with_complex_columns_resulting_into_long_names_scd"
 -- nested_stream_with_complex_columns_resulting_into_long_names from "integrationtests".test_normalization._airbyte_raw_nested_stream_with_complex_columns_resulting_into_long_names
-where _airbyte_active_row = True
+where _airbyte_active_row = 1
   );