+++ airbyte-integrations/connectors/source-bigquery/src/main/java/io/airbyte/integrations/source/bigquery/BigQuerySource.java
@@ -36,7 +36,7 @@ import io.airbyte.commons.util.AutoCloseableIterators;
 import io.airbyte.db.Databases;
 import io.airbyte.db.SqlDatabase;
 import io.airbyte.db.bigquery.BigQueryDatabase;
-import io.airbyte.db.bigquery.BigQueryUtils;
+import io.airbyte.db.bigquery.BigQuerySourceOperations;
 import io.airbyte.integrations.base.IntegrationRunner;
 import io.airbyte.integrations.base.Source;
 import io.airbyte.integrations.source.relationaldb.AbstractRelationalDbSource;
@@ -64,6 +64,7 @@ public class BigQuerySource extends AbstractRelationalDbSource<StandardSQLTypeNa
 
   private String quote = "";
   private JsonNode dbConfig;
+  private final BigQuerySourceOperations sourceOperations = new BigQuerySourceOperations();
 
   @Override
   public JsonNode toDatabaseConfig(JsonNode config) {
@@ -105,7 +106,7 @@ public class BigQuerySource extends AbstractRelationalDbSource<StandardSQLTypeNa
 
   @Override
   protected JsonSchemaPrimitive getType(StandardSQLTypeName columnType) {
-    return BigQueryUtils.getType(columnType);
+    return sourceOperations.getType(columnType);
   }
 
   @Override
@@ -114,7 +115,12 @@ public class BigQuerySource extends AbstractRelationalDbSource<StandardSQLTypeNa
   }
 
   @Override
-  protected List<TableInfo<CommonField<StandardSQLTypeName>>> discoverInternal(BigQueryDatabase database) {
+  protected List<TableInfo<CommonField<StandardSQLTypeName>>> discoverInternal(BigQueryDatabase database) throws Exception {
+    return discoverInternal(database, null);
+  }
+
+  @Override
+  protected List<TableInfo<CommonField<StandardSQLTypeName>>> discoverInternal(BigQueryDatabase database, String schema) {
     String projectId = dbConfig.get(CONFIG_PROJECT_ID).asText();
     String datasetId = getConfigDatasetId(database);
     List<Table> tables =
@@ -156,7 +162,7 @@ public class BigQuerySource extends AbstractRelationalDbSource<StandardSQLTypeNa
         enquoteIdentifierList(columnNames),
         getFullTableName(schemaName, tableName),
         cursorField),
-        BigQueryUtils.getQueryParameter(cursorFieldType, cursor));
+        sourceOperations.getQueryParameter(cursorFieldType, cursor));
   }
 
   private AutoCloseableIterator<JsonNode> queryTableWithParams(BigQueryDatabase database, String sqlQuery, QueryParameterValue... params) {