+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.1.20
+- Allow using `requests.auth.AuthBase` as authenticators instead of custom CDK authenticators.
+- Implement Oauth2Authenticator, MultipleTokenAuthenticator and TokenAuthenticator authenticators.
+- Add support for both legacy and requests native authenticator to HttpStream class.
+
+## 0.1.19
+No longer prints full config files on validation error to prevent exposing secrets to log file: https://github.com/airbytehq/airbyte/pull/5879
+
+## 0.1.18
+Fix incremental stream not saved state when internal limit config set.
+
 ## 0.1.17
 Fix mismatching between number of records actually read and number of records in logs by 1: https://github.com/airbytehq/airbyte/pull/5767